backend/requirements.txt
# CryptoFastSignals - Backend Dependencies

# Flask for API
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP requests
requests==2.31.0

# Data manipulation
pandas==2.1.4
numpy==1.26.2

# WebSocket (for future real-time updates)
python-socketio==5.10.0

# Database & Environment
supabase==2.9.0
python-dotenv==1.0.0
pytz==2023.3.post1
psutil==5.9.8

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# AI / LLM
google-generativeai>=0.3.0

# Push Notifications
pywebpush==2.2.0
cryptography>=42.0.0
